+++ src/main/interfaces/User.ts
@@ -1,0 +1,15 @@
+export interface User {
+  id: string;
+  forename: string;
+  surname: string;
+  email: string;
+  active: boolean;
+  locked: boolean;
+  pending: boolean;
+  stale: boolean;
+  roles: string[];
+  ssoProvider: string;
+  ssoId: string;
+  lastModified: string;
+  createDate: string;
+}
