+++ src/main/interfaces/UserResultData.ts
@@ -1,22 +1,7 @@
-import {PageData} from './PageData';
+import { PageData } from './PageData';
+import { User } from './User';
 
 export interface UserResultData extends PageData {
   search: string;
   result?: User;
-}
-
-export interface User {
-  id: string;
-  forename: string;
-  surname: string;
-  email: string;
-  active: boolean;
-  locked: boolean;
-  pending: boolean;
-  stale: boolean;
-  roles: string[];
-  ssoProvider: string;
-  ssoId: string;
-  lastModified: string;
-  createDate: string;
 }
